requirements.txt
aiohttp==3.9.3
beautifulsoup4==4.12.3
fastapi==0.110.0
pydantic==2.6.3
playwright==1.42.0
uvicorn==0.27.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
requests-unixsocket==0.3.0 